requirements.txt
# BioAgent Python 依赖
# 参考 POPGENAGENT 的 requirements.txt

# 核心依赖
requests>=2.28.0
pyyaml>=6.0

# Web UI
gradio>=3.0.0

# 可视化（可选）
matplotlib>=3.5.0
plotly>=5.10.0

# 知识库向量检索（可选）
# faiss-cpu>=1.7.0
# sentence-transformers>=2.2.0

# 数据处理（可选）
pandas>=1.4.0
numpy>=1.21.0

# 其他
python-dotenv>=0.19.0
